Create Absolute_Permutation.cpp
Absolute Permutation

Permutations will take place iff pos(i) is given to it's corresponding value w.r.t k and i is given to pos(i) w.r.t k, for e.g. if k = 2 and n = 4, then,
OP:  3 4 1 2
Pos: 1 2 3 4
Here, if 3 is for 1th position then for 3rd position the value can only be 1, nothing else

Cases

Case 1: n is odd and k is 0 (Possible)
Basically if n is odd and k is 0 then we need to print all natural numbers from 1 to 1
Case 2: n is odd and k > 0 (Not possible)
Case 3: n is even and k > n/2 (Not possible)
Case 4: n is even and n/k is an Odd number (Not possible because in this case one division will not have it's corresponding division)
           For e.g. n = 12, k = 4
Therefore, n/k = 3 (Odd number)
OP:  5 6 7 8 1  2 3 4 (Now, we are out of unique values for this last division. Thus, absolute permutation is not possible here.)
Pos: 1 2 3 4 5  6 7 8 9 10 11 12
Case 5: n is even and (n/k) is even (Possible)